# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BlastArena API
  version: 1.0.0
  description: |
    REST API for BlastArena, a multiplayer online grid-based explosive arena game.
    All endpoints are mounted under the `/api` prefix.
    Authentication uses JWT Bearer tokens (access token) and httpOnly cookies (refresh token).
servers:
  - url: /api
    description: API base path
...
security:
  - BearerAuth: []
paths:
...
  /messages:
    get:
      tags: [Messages]
      summary: List conversation summaries
      operationId: getConversations
      responses:
        '200':
          description: Conversation list ordered by most recent
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConversationListItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /messages/{userId}:
    get:
      tags: [Messages]
      summary: Get paginated conversation with a user
      operationId: getConversation
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/Page'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Paginated messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConversationMessage'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /messages/{userId}/read:
    put:
      tags: [Messages]
      summary: Mark messages from a user as read
      operationId: markMessagesRead
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Messages marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===================== Cosmetics =====================
...
  /maps/mine:
    get:
      tags: [Custom Maps]
      summary: List current user's maps
      operationId: getMyMaps
      responses:
        '200':
          description: User's custom maps
          content:
            application/json:
              schema:
                type: object
                properties:
                  maps:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomMapSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /maps/published:
    get:
      tags: [Custom Maps]
      summary: List all published maps
      operationId: getPublishedMaps
      responses:
        '200':
          description: Published custom maps
          content:
            application/json:
              schema:
                type: object
                properties:
                  maps:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomMapSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /maps/{id}:
    get:
      tags: [Custom Maps]
      summary: Get full map data
      description: Returns full map data if the user owns it or the map is published.
      operationId: getMap
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Full custom map data
          content:
            application/json:
              schema:
                type: object
                properties:
                  map:
                    $ref: '#/components/schemas/CustomMap'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Map not found or not accessible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Custom Maps]
      summary: Update a custom map (owner only)
      operationId: updateMap
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomMapInput'
      responses:
        '200':
          description: Map updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: Invalid map data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Map not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Custom Maps]
      summary: Delete a custom map (owner only)
      operationId: deleteMap
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Map deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Map not found or not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /maps:
    post:
      tags: [Custom Maps]
      summary: Create a new custom map
      operationId: createMap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomMapInput'
      responses:
        '201':
          description: Map created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
        '400':
          description: Invalid map data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===================== Docs =====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token passed in Authorization header
    CookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: httpOnly refresh token cookie (set by login/refresh endpoints)
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Message:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ---- Enums ----
...
    ConversationListItem:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        lastMessage:
          type: string
        lastMessageAt:
          type: string
          format: date-time
        unreadCount:
          type: integer
    ConversationMessage:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        senderUsername:
          type: string
        recipientId:
          type: integer
        message:
          type: string
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ---- Cosmetics ----
...
    CustomMap:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        mapWidth:
          type: integer
        mapHeight:
          type: integer
        tiles:
          type: array
          items:
            type: array
            items:
              type: string
        spawnPoints:
          type: array
          items:
            type: object
            properties:
              x:
                type: integer
              y:
                type: integer
        isPublished:
          type: boolean
        createdBy:
          type: integer
        creatorUsername:
          type: string
        playCount:
          type: integer
    CustomMapSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mapWidth:
          type: integer
        mapHeight:
          type: integer
        spawnCount:
          type: integer
        isPublished:
          type: boolean
        createdBy:
          type: integer
        creatorUsername:
          type: string
        playCount:
          type: integer
    CustomMapInput:
      type: object
      required: [name, mapWidth, mapHeight, tiles, spawnPoints]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
          default: ''
        mapWidth:
          type: integer
          minimum: 9
          maximum: 51
          description: Must be an odd number
        mapHeight:
          type: integer
          minimum: 9
          maximum: 51
          description: Must be an odd number
        tiles:
          type: array
          items:
            type: array
            items:
              type: string
        spawnPoints:
          type: array
          items:
            type: object
            required: [x, y]
            properties:
              x:
                type: integer
              y:
                type: integer
          minItems: 2
          maxItems: 8
        isPublished:
          type: boolean
          default: false

    # ---- Campaign Replays ----
...